+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added many unit tests to increase test coverage
+- Added an option to do a partial update of the stock
 
 ### Changed